Test work, fix user plugins disappearing

+++ kaithem/src/tests/conftest.py
@@ -4,55 +4,43 @@
 
 # It starts the whole app, pointed at a ram sandbox, to run things.
 
+import sys
 
-import builtins
+if "--collect-only" not in sys.argv:
+    import builtins
+    import os
+    import shutil
+
+    import kaithem
+    import kaithem.src.config
+
+    if os.path.exists("/dev/shm/kaithem_tests"):
+        shutil.rmtree("/dev/shm/kaithem_tests")
+
+    os.makedirs("/dev/shm/kaithem_tests/plugins/Test")
+
+    old_open = open
+
+    def open2(path, mode="r", *args, **kwargs):
+        if not (str(path).startswith("/dev/shm/")) and not path == "/dev/null":
+            if "w" in mode or "a" in mode:
+                if "__pycache__" not in str(path):
+                    raise RuntimeError("Unit testing is not allowed to write outside of /dev/shm")
+
+        return old_open(path, mode, *args, **kwargs)
+
+    builtins.open = open2
+
+    cfg = {
+        "ssl_dir": "/dev/shm/kaithem_tests/ssl",
+        "site_data_dir": "/dev/shm/kaithem_tests",
+        # Prevent it from getting IP geolocation every time
+        "location": "0,0",
+        "log_format": "normal",
+    }
+
+    pp = """
 import os
-import shutil
-
-import kaithem
-import kaithem.src.auth
-import kaithem.src.config
-
-if os.path.exists("/dev/shm/kaithem_tests"):
-    shutil.rmtree("/dev/shm/kaithem_tests")
-
-os.makedirs("/dev/shm/kaithem_tests/plugins/Test")
-
-old_open = open
-
-kaithem.src.auth.addUser("testuser", "testpass")
-kaithem.src.auth.addGroup("testgroup")
-kaithem.src.auth.addGroupPermission("testgroup", "__all_permissions__")
-kaithem.src.auth.addUserToGroup("testuser", "testgroup")
-
-
-def open2(path, mode="r", *args, **kwargs):
-    if not (str(path).startswith("/dev/shm/")):
-        if "w" in mode or "a" in mode:
-            if "__pycache__" not in str(path):
-                raise RuntimeError("Unit testing is not allowed to write outside of /dev/shm")
-
-    return old_open(path, mode, *args, **kwargs)
-
-
-builtins.open = open2
-
-cfg = {
-    "ssl_dir": "/dev/shm/kaithem_tests/ssl",
-    "site_data_dir": "/dev/shm/kaithem_tests",
-    # Prevent it from getting IP geolocation every time
-    "location": "0,0",
-    "log_format": "normal",
-}
-
-
-pp = """
-import os
-import kaithem.api.web as webapi
-
-
-tp = os.path.join(os.path.dirname(__file__), "template.html")
-
 
 import kaithem.api.chandler as chandlerapi
 
@@ -67,9 +55,9 @@ def foo_command(x: str):
     return True
 
 chandlerapi.add_command("foo_command", foo_command)
-"""
+    """
 
-pt = """
+    pt = """
 {% extends "pagetemplate.j2.html" %}
 
 {% block title %}Title Here{% endblock %}
@@ -89,10 +77,9 @@ pt = """
     </section>
 </main>
 {% endblock %}
-"""
+    """
 
-
-test_scene = """
+    test_scene = """
 active: true
 alpha: 1
 backtrack: true
@@ -102,19 +89,19 @@ bpm: 60
 command_tag: ''
 crossfade: 0
 cues:
-  c1:
+c1:
     length: 0.5
     number: 10000
     track: false
     values:
-      /unit_testing/t2:
+    /unit_testing/t2:
         value: 183.0
 
-  default:
+default:
     length: 0.5
     number: 5000
     values:
-      /unit_testing/t1:
+    /unit_testing/t1:
         value: 132.0
 default_next: ''
 display_tags: []
@@ -129,61 +116,61 @@ notes: ''
 priority: 50
 slide_overlay_url: ''
 slideshow_layout: |
-  <style>
-      slideshow-app {
-          display: flex;
-          flex-wrap: wrap;
-          flex-direction: row;
-      }
+<style>
+    slideshow-app {
+        display: flex;
+        flex-wrap: wrap;
+        flex-direction: row;
+    }
 
-      main{
+    main{
         display: flex;
         flex-direction: column;
         flex-grow:10;
-      }
+    }
 
-      media-player {
-          flex-grow: 5;
-      }
+    media-player {
+        flex-grow: 5;
+    }
 
-      .sidebar {
-          background: linear-gradient(175deg, rgba(36,36,36,1) 0%, rgba(77,77,77,1) 100%);
-          max-width: calc(max(30%, min(24em, 100%) ));
-          text-wrap: wrap;
-      }
+    .sidebar {
+        background: linear-gradient(175deg, rgba(36,36,36,1) 0%, rgba(77,77,77,1) 100%);
+        max-width: calc(max(30%, min(24em, 100%) ));
+        text-wrap: wrap;
+    }
 
-  </style>
+</style>
 
 
-  <slideshow-app>
-      <main>
-          <header></header>
-          <media-player></media-player>
-          <footer></footer>
-      </main>
-      <div class="sidebar" v-if="cueText" v-html="cueText">
-      </div>
-  </slideshow-app>
+<slideshow-app>
+    <main>
+        <header></header>
+        <media-player></media-player>
+        <footer></footer>
+    </main>
+    <div class="sidebar" v-if="cueText" v-html="cueText">
+    </div>
+</slideshow-app>
 sound_output: ''
 utility: false
 uuid: efcae37b3e78437cad5098eadf3a172d
 
-"""
+    """
 
-with open("/dev/shm/kaithem_tests/plugins/Test/__init__.py", "w") as f:
-    f.write(pp)
+    with open("/dev/shm/kaithem_tests/plugins/Test/__init__.py", "w") as f:
+        f.write(pp)
 
-with open("/dev/shm/kaithem_tests/plugins/Test/template.html", "w") as f:
-    f.write(pt)
+    with open("/dev/shm/kaithem_tests/plugins/Test/template.html", "w") as f:
+        f.write(pt)
 
-os.makedirs("/dev/shm/kaithem_tests/chandler/scenes")
+    os.makedirs("/dev/shm/kaithem_tests/chandler/scenes")
 
-with open("/dev/shm/kaithem_tests/chandler/scenes/unit_testing.yaml", "w") as f:
-    f.write(test_scene)
+    with open("/dev/shm/kaithem_tests/chandler/scenes/unit_testing.yaml", "w") as f:
+        f.write(test_scene)
 
-kaithem.initialize_app(cfg)
-# TODO Sound can't be imported before config init, eventually
-# A refactor should fix that
-import kaithem.src.sound  # noqa
+    kaithem.initialize_app(cfg)
+    # TODO Sound can't be imported before config init, eventually
+    # A refactor should fix that
+    import kaithem.src.sound  # noqa
 
-kaithem.src.sound.select_backend("test")
+    kaithem.src.sound.select_backend("test")
